move the start and shutdown system to api_utils.py
