+++ examples/hello-world/hw_pre_start.sh
@@ -23,7 +23,7 @@ nvflare poc --start -ex admin
 
 # Check if the FL system is ready
 python <<END
-from nvflare.lighter.utils import wait_for_system_start
+from nvflare.tool.api_utils import wait_for_system_start
 wait_for_system_start(num_clients = 2,  prod_dir = "${NVFLARE_POC_WORKSPACE}")
 
 END
